+++ src/rsnapshot_docker_compose_backup/config/backup.ini
@@ -1,23 +1,22 @@
-
 #This is the default config of the rsnapshot-docker-compose-backup program.
 
 #This section can contain rsnapshot commands that are executed in the different steps of the container backup
-[default_config]
+[{default_config}]
 #Example:
 #run_backup=backup_script   echo "Hello" > helloWorld   test/
 
 #This section contains settings that can be set. This section can only be used in the default config.
-[settings]
+[{default_config_settings}]
 #This outputs the Start and End Times for each backup command to the log. 
 #Rsnapshot doesn't log timestamps
 logTime = true
 
-[default_config.vars]
+[{default_config_vars}]
 #This setting corresponds to the var with the same name and can be used as a prefix in the folder path
 backupprefixfolder = .
 
 #This Section controls which actions should be enabled
-[default_config.actions]
+[{default_config_actions}]
 #This action stops the container before copying the volumes and starts it after it finished.
 stopcontainer = true
 
@@ -33,21 +32,22 @@ logbackup = false
 projectDirBackup = false
 
 #The following is the definition of actions that can be used in the backup
-[actions.volumeBackup]
-backup = backup	$volumes.path	$backupPrefixFolder/$serviceName/$volumes.name
 
-[actions.yamlBackup]
-runtime_backup = backup	$projectFolder	$backupPrefixFolder/$serviceName/yaml	+rsync_long_args=--include=*.yml,+rsync_long_args=--include=*.yaml
+[{actions}.volumeBackup]
+backup = backup\t$volumes.path\t$backupPrefixFolder/$serviceName/$volumes.name
 
-[actions.projectDirBackup]
-backup = backup	$projectFolder	$backupPrefixFolder/$serviceName/projectDir
+[{actions}.yamlBackup]
+runtime_backup = backup\t$projectFolder\t$backupPrefixFolder/$serviceName/yaml\t+rsync_long_args=--include=*.yml,+rsync_long_args=--include=*.yaml
 
-[actions.imageBackup]
-runtime_backup = backup_script	/usr/bin/docker image save $image -o $serviceName_image.tar	$backupPrefixFolder/$serviceName/image
+[{actions}.projectDirBackup]
+backup = backup\t$projectFolder\t$backupPrefixFolder/$serviceName/projectDir
 
-[actions.logBackup]
-backup = backup_script	/usr/bin/docker logs $containerID > $serviceName_logs.log 2>&1	$backupPrefixFolder/$serviceName/log
+[{actions}.imageBackup]
+runtime_backup = backup_script\t/usr/bin/docker image save $image -o $serviceName_image.tar\t$backupPrefixFolder/$serviceName/image
 
-[actions.stopContainer]
-stop = backup_exec	cd $projectFolder; /usr/bin/docker-compose stop
-restart = backup_exec	cd $projectFolder; /usr/bin/docker-compose start
+[{actions}.logBackup]
+backup = backup_script\t/usr/bin/docker logs $containerID > $serviceName_logs.log 2>&1\t$backupPrefixFolder/$serviceName/log
+
+[{actions}.stopContainer]
+stop = backup_exec\tcd $projectFolder; /usr/bin/docker-compose stop
+restart = backup_exec\tcd $projectFolder; /usr/bin/docker-compose start
