+++ src/rsnapshot_docker_compose_backup/structure/project.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+from pathlib import Path
+
+
+@dataclass
+class Project:
+    name: str
+    folder: Path
